backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0